--- a/README.md
+++ b/README.md
@@ -18,7 +18,7 @@ This module just makes working with multihashes a bit nicer.
 [js-multihash](//github.com/jbenet/js-multihash) is only for
 encoding/decoding multihashes, and does not depend on other libs.
 This module will depend on various implementations for each hash.
-For now, it just uses `crypto`, but will use `sha3` and `blake2`, etc.
+It currently uses `crypto` and [`sha3`](https://github.com/phusion/node-sha3) in Node.js. In the browser [`webcrypto`](https://developer.mozilla.org/en-US/docs/Web/API/SubtleCrypto) and [`browserify-sha3`](https://github.com/wanderer/browserify-sha3) are used.
 
 ## Table of Contents
 
@@ -80,16 +80,20 @@ You will need to use Node.js `Buffer` API compatible, if you are running inside 
 var multihashing = require('multihashing')
 var buf = new Buffer('beep boop')
 
-// by default returns a multihash.
-multihashing(buf, 'sha1')
+multihashing(buf, 'sha1', function (err, multihash) {
+  // by default calls back with a multihash.
+})
 
 // Use `.digest(...)` if you want only the hash digest (drops the prefix indicating the hash type).
-multihashing.digest(buf, 'sha1')
+multihashing.digest(buf, 'sha1', function (err , digest) {
+  // digest is the raw digest
+})
 
-// Use `.createHash(...)` for a `crypto.createHash` interface.
+// Use `.createHash(...)` for the raw hash functions
 var h = multihashing.createHash('sha1')
-h.update(buf)
-h.digest()
+h(buf, (err, digest) => {
+  // digest is a buffer of the sha1 of buf
+})
 ```
 
 ## Examples
@@ -100,39 +104,23 @@ h.digest()
 > var multihashing = require('multihashing')
 > var buf = new Buffer('beep boop')
 
-> console.log(multihashing(buf, 'sha1'))
+> multihashing(buf, 'sha1'), function (err, mh) { console.log(mh) })
 // => <Buffer 11 14 7c 83 57 57 7f 51 d4 f0 a8 d3 93 aa 1a aa fb 28 86 3d 94 21>
 
-> console.log(multihashing(buf, 'sha2-256'))
+> multihashing(buf, 'sha2-256', function (err, mh) { console.log(mh) })
 // => <Buffer 12 20 90 ea 68 8e 27 5d 58 05 67 32 50 32 49 2b 59 7b c7 72 21 c6 24 93 e7 63 30 b8 5d dd a1 91 ef 7c>
 
-> console.log(multihashing(buf, 'sha2-512'))
+> multihashing(buf, 'sha2-512'), function (err, mh) { console.log(mh) })
 // => <Buffer 13 40 14 f3 01 f3 1b e2 43 f3 4c 56 68 93 78 83 77 1f a3 81 00 2f 1a aa 5f 31 b3 f7 8e 50 0b 66 ff 2f 4f 8e a5 e3 c9 f5 a6 1b d0 73 e2 45 2c 48 04 84 b0 ...>
-```
-
-### Raw digest output
-
-```js
-> var multihashing = require('multihashing')
-> var buf = new Buffer('beep boop')
-
-> console.log(multihashing.digest(buf, 'sha1'))
-// => <SlowBuffer 7c 83 57 57 7f 51 d4 f0 a8 d3 93 aa 1a aa fb 28 86 3d 94 21>
-
-> console.log(multihashing.digest(buf, 'sha2-256'))
-// => <SlowBuffer 90 ea 68 8e 27 5d 58 05 67 32 50 32 49 2b 59 7b c7 72 21 c6 24 93 e7 63 30 b8 5d dd a1 91 ef 7c>
-
-> console.log(multihashing.digest(buf, 'sha2-512'))
-// => <SlowBuffer 14 f3 01 f3 1b e2 43 f3 4c 56 68 93 78 83 77 1f a3 81 00 2f 1a aa 5f 31 b3 f7 8e 50 0b 66 ff 2f 4f 8e a5 e3 c9 f5 a6 1b d0 73 e2 45 2c 48 04 84 b0 2e 03 ...>
 ```
 
 ## API
 
-### `multihashing(buf, func, length)`
+### `multihashing(buf, func, [length,] callback)`
 
-### `digest(buf, func, length)`
+### `digest(buf, func, [length,] callback)`
 
-### `createHash(func, length)`
+### `createHash(func)`
 
 ### `functions`