In Node.js the underlying hashing functions have not changed, but the browser now uses `webcrypto` instead of JavaScript based methods for `SHA1`, `SHA2-256` and `SHA2-512`.

Also `SHA3` support was added in both Node.js and the browser.

BREAKING CHANGE:

The api was changed to be callback based, as webcrypto only exposes async methods.

Closes #10


## Size Impact

|            | non-minified | minified |
|-------|----------|-----|
|before|`953K` | `620K`|
|after   | `328K` | `208K`|

## Speed Impact

|               | Node.js 6.6.0  | Browser Chrome |
|---------------|----------------|----------------|
| before tiny   |                |                |
|   sha1        | 69,814 ops/sec | 15,976 ops/sec |
|   sha256      | 64,766 ops/sec | 10,957 ops/sec |
|   sha512      | 56,372 ops/sec | 17,180 ops/sec |
| after tiny    |                |                |
|   sha1        | 52,057 ops/sec |  5,701 ops/sec |
|   sha256      | 51,105 ops/sec |  5,870 ops/sec |
|   sha512      | 48,676 ops/sec |  5,846 ops/sec |
| before medium |                |                |
|   sha1        | 27,808 ops/sec |  2,908 ops/sec |
|   sha256      | 17,605 ops/sec |  2,962 ops/sec |
|   sha512      | 22,641 ops/sec |    852 ops/sec |
| after medium  |                |                |
|   sha1        | 19,619 ops/sec |  3,862 ops/sec |
|   sha256      | 14,980 ops/sec |  3,186 ops/sec |
|   sha512      | 17,420 ops/sec |  3,256 ops/sec |
| before large  |                |                |
|   sha1        |  4,330 ops/sec |    308 ops/sec |
|   sha256      |  2,497 ops/sec |    323 ops/sec |
|   sha512      |  3,407 ops/sec |     88 ops/sec |
| after large   |                |                |
|   sha1        |  4,192 ops/sec |  1,145 ops/sec |
|   sha256      |  2,472 ops/sec |    899 ops/sec |
|   sha512      |  3,382 ops/sec |  1,070 ops/sec |


- tiny: `Buffer(100)`
- medium: `Buffer(10 * 1024)`
- large: `Buffer(100 * 1024)`


Due to the overhead, on very small buffers the pure JavaScript performs better. When looking at larger inputs it comes apparent that WebCrypto is a good deal faster.