@diasdavid 2016-09-22
Cool!

Any good reason why to break the interface? We can support a callback API and a sync API at the same time.

Can you add benchmarking tests to justify/motivate this change

PS: Can we get that coveralls comments off?