@dignifiedquire 2016-09-22
The dependency size reduction is massive, I will post numbers later. 

Speed wise I haven't tested it, but I expect good things. 

We can't have a sync interface on the browser because of webcrypto as we have seen before with my work libp2p-crypto and it would be very confusing when exposing one interface in node and another in the browser. 

I don't have the rights to turn the off on coveralls for some reason.