@dignifiedquire 2016-09-25
@diasdavid benchmarks are there, plus full coverage is back