@diasdavid 2016-09-30
@dignifiedquire this is really, really great :D having a faster js-ipfs and libp2p (specially for testing, which means developer produtivity is awesome).

However, as I mentioned in IRC, I don't feel confident of breaking the multihashing API, instead, let's create a `js-multihashing-async` and reference it from this module as being a faster, smaller version.

For those of you reading this PR **"Why create two modules and not have the two API in one?"**, because having the sync version would mean that we have to keep all the dependencies for the sync code to work, keeping the module really big. It is true that now with things like rollup, you can eliminate dead branches, but also means everyone has to understand and buy into that.