@dignifiedquire 2016-09-30
> However, as I mentioned in IRC, I don't feel confident of breaking the multihashing API, instead, let's create a js-multihashing-async and reference it from this module as being a faster, smaller version.

I still don't understand why it is not enough to just bump the semver major version. If we create a second module we have to maintain it + backport fixes to this module including things like sha3. So this means a lot more work in the long time. In addition if others are still using this module and we use the async module in our code base others would have to bundle both versions into their code even increasing their bundles in size.