@diasdavid 2016-10-01
> I still don't understand why it is not enough to just bump the semver major version

multihashing and the other multiformats implementations should not be seen as a 'internal' of IPFS as they are used by a lot of other groups, that being said, we can't just break user space like that.

> If we create a second module we have to maintain it + backport fixes to this module including things like sha3

multihashing is incredibly small and straight forward compared to other modules we maintain, checking up the [contribution graph](https://github.com/multiformats/js-multihashing/graphs/contributors), we can see that most changes have been of code refactoring or dependencies update, it is fairly stable at this point.

> In addition if others are still using this module and we use the async module in our code base others would have to bundle both versions into their code even increasing their bundles in size.

That is why documentation (examples, notes on the README) are king, informing our users that we've moved to multihashing-async because of X, Y and Z is more valuable than just flipping the API and implementation under the carpet.